% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file_cran.R
\name{cran_archive_file}
\alias{cran_archive_file}
\title{Read a file from CRAN}
\usage{
cran_archive_file(pac, version, file, repos = "https://cran.rstudio.com/")
}
\arguments{
\item{pac}{\code{character} package name.}

\item{version}{\code{character} package version.}

\item{file}{\code{character} file name to read. Possible values are \code{DESCRIPTION} and \code{NAMESPACE}.}

\item{repos}{\code{character} vector repositories URLs to use. Used only for the validation. Default \verb{https://cran.rstudio.com/}}
}
\description{
Read a file from CRAN package source.
}
\keyword{internal}
