% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file_github.R
\name{read_github_file}
\alias{read_github_file}
\title{Read a file from a GitHub CRAN repository}
\usage{
read_github_file(pac, version, file, repos = "https://cran.rstudio.com/")
}
\arguments{
\item{pac}{\code{character} package name.}

\item{version}{\code{character} package version.}

\item{file}{\code{character} file name to read. Possible values are \code{DESCRIPTION} and \code{NAMESPACE}.}

\item{repos}{\code{character} vector repositories URLs to use. Used only for the validation. Default \verb{https://cran.rstudio.com/}}
}
\description{
Read a file from a GitHub CRAN repository.
}
\note{
if the file is not found in the GitHub repository, it will try to find it in the CRAN archive.
}
\keyword{internal}
